pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aimodel-cli"
version = "1.0.0"
description = "Command-line interface for AI model management and download"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.25.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "Pillow>=8.0.0",
    "send2trash>=1.8.0",
    "beautifulsoup4>=4.9.0",
    "pysocks>=1.7.0",
    "packaging>=20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.812",
]

[project.scripts]
aimodel = "aimodel_cli.main:cli"

[project.urls]
Homepage = "https://github.com/shingo1228/aimodel-cli"
Repository = "https://github.com/shingo1228/aimodel-cli"
Issues = "https://github.com/shingo1228/aimodel-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true